requirements.txt
discord.py==2.3.2
openai==0.28.0
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0 